+++ app/code/Bulbulatory/Recommendations/Model/Source/Status.php
@@ -10,6 +10,8 @@ use Magento\Framework\Option\ArrayInterface;
  */
 class Status implements ArrayInterface
 {
+    const STATUS_UNCONFIRMED = 'Unconfirmed';
+    const STATUS_CONFIRMED = 'Confirmed';
 
     /**
      * @return array|array[]
@@ -17,8 +19,8 @@ class Status implements ArrayInterface
     public function toOptionArray()
     {
         return [
-            ['value' => 0, 'label' => __('Unconfirmed')],
-            ['value' => 1, 'label' => __('Confirmed')]
+            ['value' => 0, 'label' => __(self::STATUS_UNCONFIRMED)],
+            ['value' => 1, 'label' => __(self::STATUS_CONFIRMED)]
         ];
     }
 }